+++ domain/src/main/java/pt/uminho/dss/restaurante/domain/entity/Categoria.java
@@ -1,7 +1,7 @@
 package pt.uminho.dss.restaurante.domain.entity;
 
 public class Categoria {
-    public Long id;
+    public Integer id;
     public String nome;
     public boolean ativa;
 
@@ -14,8 +14,8 @@ public class Categoria {
 
     // Getters e Setters
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public String getNome() { return nome; }
     public void setNome(String nome) { this.nome = nome; }
